try with full retry logic

+++ src/scripts/yarn/run_command.sh
@@ -4,9 +4,17 @@
 trap 'echo "fail detected"; touch /tmp/failure' ERR
 
 echo "Running command: ${COMMAND:?}"
-bash -c "for _ in {0..${MAX_RETRIES:?}}; do ${COMMAND:?} && break || sleep ${SLEEP_TIME:?} && echo \"Retrying $COMMAND\"; done"
+for _ in $(seq 0 "${MAX_RETRY:?}"); do
+    if bash -c "${COMMAND?}"; then
+        if (( ${SHOULD_REMOVE_LOCKFILE?} )); then
+            echo "Removing lock file ${LOCK_FILE?}" 
+            rm -rf "${LOCK_FILE}"
+        fi
+        exit 0
+    fi
+    sleep "${SLEEP:?}"
+    echo "Retrying command: ${COMMAND?}"
+done
 
-if (( ${SHOULD_REMOVE_LOCKFILE?} )); then
-    echo "Removing lock file ${LOCK_FILE?}" 
-    rm -rf "${LOCK_FILE}"
-fi
+echo "failed: ${COMMAND?}" >&2
+exit 1
